+++ single_channel_player_app/windows/flutter/generated_plugins.cmake
@@ -4,6 +4,7 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   connectivity_plus
+  media_kit_video
   screen_brightness_windows
   url_launcher_windows
   volume_controller